Add script to oen Zeal if not focused
I like to have a keyboard shortcut configured to open Zeal
with Meta+z.
I like to open this in whichever workspace I'm in, and if it is
already open, to minimize it.
I also want it to be started if not started already.

TODO: Add a KDE shortcut scheme with the shortcut.