Add llvm install script
Note: This will need to be updated with a new release
of LLVM in the future.

One way to avoid this is to use
```bash
sudo bash -c "$(wget -O - https://apt.llvm.org/llvm.sh)"
```
However, the script asks for permission while adding repo.
I want to change this.