Add option to disable clang format on save if needed
Some codebases started using clang-format later in the lifecycle
so they only apply clang formatting rules to modifications.

I need to get this working. But, in the meantime, I use this option
to disable clang format on save where this is the case.

TODO: Figure out config for clang format on save